+++ apps/web/src/assets/images/diary-ava.png
