+++ apps/web/src/views/Settings/Contacts.tsx
@@ -1,6 +1,6 @@
 import { Icon28Users } from '@vkontakte/icons'
 import { Avatar, Group, Header, Link, SimpleCell } from '@vkontakte/vkui'
-import { diaryAva, winxAva } from '../../images/config.ts'
+import { diaryAva, winxAva } from '../../config/images.ts'
 
 const Contacts = () => {
   return (
